moved socket listeners
